requirements.txt
keras==3.2.1
matplotlib==3.7.2
numpy==1.25.1
opencv_python==4.8.1.78
pandas==2.0.3
Pillow==10.0.0
Pillow==10.3.0
scikit_learn==1.3.1
seaborn==0.13.2
streamlit==1.32.2
tensorflow==2.16.1
tensorflow_intel==2.16.1
